Re-arrange test files
